chore: consolidate StorageLike into src/utils/storage
